% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModeDissonanceSantiago.R
\docType{package}
\name{ModeDissonanceSantiago}
\alias{ModeDissonanceSantiago}
\alias{ModeDissonanceSantiago-package}
\title{ModeDissonanceSantiago}
\description{
The ModeDissonanceSantiago is a companion package for **paper**. The package includes
  the data used in the analysis of mode dissonance in Santiago, Chile; it also includes
  one utility function for cross-tabulations.
}
\author{
Beatriz G. Mella Lira, \email{beatriz.lira.14@ucl.ac.uku}

Antonio Paez, \email{paezha@mcmaster.ca}
}
